Completed the second phase of the code grouping. Now each group of code resides in its own source file with a meaninful name. We no longer have an uber AnalysisGraph.cpp source file
