+++ lib/printing.cpp
@@ -1,0 +1,99 @@
+#include "AnalysisGraph.hpp"
+#include <boost/range/algorithm/for_each.hpp>
+
+using namespace std;
+using fmt::print;
+
+
+/*
+ ============================================================================
+ Public: Printing
+ ============================================================================
+*/
+
+void AnalysisGraph::print_nodes() {
+  print("Vertex IDs and their names in the CAG\n");
+  print("Vertex ID : Name\n");
+  print("--------- : ----\n");
+  for_each(this->node_indices(), [&](int v) {
+    cout << v << "         " << this->graph[v].name << endl;
+  });
+}
+
+void AnalysisGraph::print_edges() {
+  for_each(edges(), [&](auto e) {
+    cout << "(" << (*this)[boost::source(e, this->graph)].name << ", "
+         << (*this)[boost::target(e, this->graph)].name << ")" << endl;
+  });
+}
+
+void AnalysisGraph::print_name_to_vertex() {
+  for (auto [name, vert] : this->name_to_vertex) {
+    cout << name << " -> " << vert << endl;
+  }
+  cout << endl;
+}
+
+void AnalysisGraph::print_indicators() {
+  for (int v : this->node_indices()) {
+    cout << v << ":" << (*this)[v].name << endl;
+    for (auto [name, vert] : (*this)[v].nameToIndexMap) {
+      cout << "\t"
+           << "indicator " << vert << ": " << name << endl;
+    }
+  }
+}
+
+void AnalysisGraph::print_A_beta_factors() {
+  int num_verts = this->num_vertices();
+
+  for (int row = 0; row < num_verts; ++row) {
+    for (int col = 0; col < num_verts; ++col) {
+      cout << endl << "Printing cell: (" << row << ", " << col << ") " << endl;
+      if (this->A_beta_factors[row][col]) {
+        this->A_beta_factors[row][col]->print_beta2product();
+      }
+    }
+  }
+}
+
+void AnalysisGraph::print_all_paths() {
+  int num_verts = this->num_vertices();
+
+  if (this->A_beta_factors.size() != num_verts ||
+      this->A_beta_factors[0].size() != num_verts) {
+    this->find_all_paths();
+  }
+
+  cout << "All the simple paths of:" << endl;
+
+  for (int row = 0; row < num_verts; ++row) {
+    for (int col = 0; col < num_verts; ++col) {
+      if (this->A_beta_factors[row][col]) {
+        this->A_beta_factors[row][col]->print_paths();
+      }
+    }
+  }
+}
+
+// Given an edge (source, target vertex ids - i.e. a β ≡ ∂target/∂source),
+// print all the transition matrix cells that are dependent on it.
+void AnalysisGraph::print_cells_affected_by_beta(int source, int target) {
+  typedef multimap<pair<int, int>, pair<int, int>>::iterator MMapIterator;
+
+  pair<int, int> beta = make_pair(source, target);
+
+  pair<MMapIterator, MMapIterator> beta_dept_cells =
+      this->beta2cell.equal_range(beta);
+
+  cout << endl
+       << "Cells of A afected by beta_(" << source << ", " << target << ")"
+       << endl;
+
+  for (MMapIterator it = beta_dept_cells.first; it != beta_dept_cells.second;
+       it++) {
+    cout << "(" << it->second.first * 2 << ", " << it->second.second * 2 + 1
+         << ") ";
+  }
+  cout << endl;
+}
